move workers to background-worker and remove the handlers
